SDCS-X/vOT-Management/Automated Dev/openplc-mqtt-publisher/requirements.txt
paho-mqtt
influxdb-client
requests
